Update to be compatible with MM-IoT-SDK release 2.7.2 for ESP32
In previous MM-IoT-SDK releases the regulatory database was defined in
`framework/morselib/include/mmwlan_regdb.def` and that could just be
included. However, as of this release the end application needs to
provide the regulatory database.

This change brings the reference regulatory database provided by Morse
Micro into the halow component.