Add Gitea actions to test builds and formatting
This change adds some basic Gitea actions to validate that the code
builds and is formatted correctly. May want to consider adding our
own docker images to avoid the install steps but the overhead is
fairly low for now.

Additionally, added a `format_check` target that can be used to
validate the formatting.